refactor(trusty-agents-common): extract skills deploy/manifest/tiers machinery from trusty-mpm (refs #2892, #2818)
Moves the skill deploy pipeline (skill_deployer), the checksum +
atomic-write ownership manifest (skill_manifest), and the
project-custom/user-custom/bundled precedence resolver (skill_tiers) out
of trusty-mpm's binary crate into a new trusty-agents-common::skills
module, mirroring the agent compose/deploy/manifest extraction (#2892,
PR #2909). trusty-mpm re-exports every item from its existing
core::skill_deployer / core::skill_manifest / core::skill_tiers module
paths for source compatibility, so every call site compiles unchanged.

skills::manifest reuses agents::manifest's self-contained ManifestError
(thiserror) and its atomic_write/checksum helpers rather than defining a
second error type — the skill manifest needs exactly the same two
failure modes (IO + JSON) the agent manifest already carries.
skills::deployer::is_skill_file/skill_stem widen from pub(crate) to pub
(mirroring agents::deployer::is_agent_file) so trusty-mpm's
skill_staleness module can keep calling them cross-crate.

skills::tiers.rs splits its test module into tiers_tests.rs (mirroring
the builder/builder_tests split) to stay under the 500-line SLOC
production cap.

Repoints the grandfathered skill_manifest_save_is_atomic dangling
Test: pointer in .test-pointer-allowlist.tsv to the new
crates/trusty-agents-common/src/skills/manifest.rs path.

This unblocks trusty-code consuming the same skill catalog machinery
instead of forking it (#2892).
